pyproject.toml
[project]
name = "mono-log-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "boto3>=1.42.38",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "hypercorn>=0.18.0",
    "orjson>=3.11.5",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.1.0",
    "pillow-avif-plugin>=1.5.5",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "redis>=7.1.0",
    "ruff>=0.14.14",
    "sqlalchemy>=2.0.46",
    "ty>=0.0.14",
    "typer>=0.21.1",
    "uuid-extension>=0.2.0",
    "uvicorn>=0.40.0",
]
